requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
selenium>=4.15
supabase_py>=1.0

